Add HostProperty enum
In addition to tunables (which tweak how the computer works) and
capabilities (which tell us what jobs the computer runs), we now
add HostProperty, which tells us specific things about the computer.

There is probably a lot of overlap between tunables and properties,
so use the Potter "I'll know it when I see it" test when deciding
if something is a tunable or a property.