Use configinit to set up containers
Leverage firstboot_sentinel and frieze_configinit scripts to
intialize containers when they are restarted.

Configinit payload delivery is achieved as follows: when a container
is configured, all capabilities within that container are used to
generate a tarball, which is then base64 encoded and pushed to
the pickup point (/root/cfinit) in the jail just like any other file.

frieze_configinit has been modified to check whether it is in a jail;
if it is, it checks the pickup point for the config tarball, decodes it,
and then applies it. If it is on a regular bare-metal host, it still
checks the link-local address for the configuration payload.