DNS-based load balancing for services
When adding capabilities to a deployment, allow specification of
stripe_group parameter which maps to a group of identical containered
services to a single alias. This alias is entered in dns, which then
distributes incoming requests to it in round-robin fashion.

Also introduce a pflate rc script depending on named to ensure that
pf can always resolve aliases.