Add user roles and allow their assignment to capabilities
Roles represent users of capabilities and are unique in a domain
context.

They can be used to restrict access to a capability by passing them
to add_capability via the acls named parameter. If acls is empty, the
capability is notionally empty; if acls is set, access should be
restricted to the users in the list.

Keep in mind that while frieze can make the credentails available to
a capability in a standardized fashion, the underlying software must
*still* be set to present and authenticate them.